requirements.txt
appdirs==1.4.3
asgiref==3.2.7
beautifulsoup4==4.9.1
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coverage==5.2.1
distlib==0.3.0
distro==1.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.7
django-adminlte2==0.4.3
django-bootstrap4==2.1.1
django-heroku==0.3.1
django-stdimage==5.1.1
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
ipaddr==2.2.0
lockfile==0.12.2
model-mommy==2.0.0
msgpack==0.6.2
mysql==0.0.2
mysqlclient==2.0.1
packaging==20.3
pep517==0.8.2
Pillow==7.1.2
progress==1.5
progressbar2==3.51.4
psycopg2==2.8.5
psycopg2-binary==2.8.5
PyMySQL==0.9.3
pyparsing==2.4.7
python-utils==2.4.0
pytoml==0.1.21
pytz==2020.1
requests==2.23.0
retrying==1.3.3
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
static3==0.7.0
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.1.0
